backend/requirements.txt
numpy
utils
Flask
SQLAlchemy
gunicorn
flask
flask-cors
python-dotenv
pandas
requests
sqlalchemy
psycopg2-binary
kagglehub
beautifulsoup4
lxml